+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - #2258 remember saved position in key map list when changing them. Do not flash with loading indicator briefly as well.
 - #2167 do not show errors for app not in foreground constraint when the app is uninstalled/disabled.
 - #2262 request local network access permission for HTTP request action on Android 17+.
+- #2261 show the app's saved name instead of a generic label or raw package name for app-related constraints and actions when the app has been uninstalled.
 
 ## [4.4.0](https://github.com/sds100/KeyMapper/releases/tag/v4.4.0)